backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
python-dotenv==1.0.1
PyYAML==6.0.2
mangum==0.21.0
